services/metrics-worker/requirements.txt
fastapi==0.111.0
psycopg2-binary==2.9.9
uvicorn[standard]==0.30.1
